Fix various issues with expression handling.
Expressions, with and without parenthesis, currently do not behave
properly, and exhibit an unwanted nesting behavior. Change their
definition in `Parser.jjt`, and rewrite their handling in `RuntimeNode`
in order to remedy these problems.

Also, make a change to `RuntimeContext` to fix a simple bug where
multiple conditions would be matched in `updateObject`.